backend/backend/requirements.txt
asgiref==3.9.1
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
google-genai>=0.3.0
pillow==11.1.0
chromadb
pypdf
python-docx
beautifulsoup4
youtube-transcript-api
weasyprint
python-pptx
openpyxl
trafilatura
pymupdf4llm
yt-dlp
sendgrid
django-ratelimit
pydub
audioop-lts; python_version >= '3.13'
